ENH: move SiriusHexaSpinbox to widgets subpackage
